WIP Vector_get size_t

+++ AffinityPanel.c
@@ -163,7 +163,7 @@ static void AffinityPanel_updateTopo(AffinityPanel* this, MaskItem* item) {
    if (item->sub_tree == 2)
       return;
 
-   for (int i = 0; i < Vector_size(item->children); i++)
+   for (size_t i = 0; i < Vector_size(item->children); i++)
       AffinityPanel_updateTopo(this, (MaskItem*) Vector_get(item->children, i));
 }
 
@@ -181,7 +181,7 @@ static void AffinityPanel_update(AffinityPanel* this, bool keepSelected) {
    if (this->topoView) {
       AffinityPanel_updateTopo(this, this->topoRoot);
    } else {
-      for (int i = 0; i < Vector_size(this->cpuids); i++) {
+      for (size_t i = 0; i < Vector_size(this->cpuids); i++) {
          AffinityPanel_updateItem(this, (MaskItem*) Vector_get(this->cpuids, i));
       }
    }
@@ -444,7 +444,7 @@ Affinity* AffinityPanel_getAffinity(Panel* super, Machine* host) {
       Affinity_add(affinity, (unsigned)i);
    hwloc_bitmap_foreach_end();
    #else
-   for (int i = 0; i < Vector_size(this->cpuids); i++) {
+   for (size_t i = 0; i < Vector_size(this->cpuids); i++) {
       const MaskItem* item = (const MaskItem*)Vector_get(this->cpuids, i);
       if (item->value) {
          Affinity_add(affinity, item->cpu);
